Moved code for backup-cdsp to offline.py
